backend/django_app/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.5
python-dotenv>=0.21.0
minio>=7.1.14
django-cors-headers>=3.13.0
python-dateutil>=2.8.2
Pillow>=9.0.0
djangorestframework-simplejwt>=5.2.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
